MovieFlipCard/MovieCardOne Refactor
There was a depriciated component: MovieFlipCard, which was removed. MovieCardOne has now been renamed MovieDetailsCard
