+++ src/components/MovieDetailsCard/MovieDetailsCard.styles.ts
