+++ src/screens/SearchScreen/SearchScreen.tsx
@@ -15,7 +15,7 @@ import { searchMovies, getMovieDetails } from "@/src/utils/APIs/TMDB";
 import SearchItem from "@/src/components/SearchItem";
 import { tmdbMovie, DjangoMovie } from "@/src/utils/types/types";
 import { getMyMovie } from "@/src/utils/APIs/api";
-import MovieCardOne from "@/src/components/MovieFlipCard/MovieCardOne";
+import MovieDetailsCard from "@/src/components/MovieDetailsCard/MovieDetailsCard";
 
 const SearchScreen = () => {
   // Get the current theme colors from the navigation context
@@ -161,7 +161,7 @@ const SearchScreen = () => {
               {/* Prevent closing modal when tapping inside the modal content */}
               <TouchableWithoutFeedback>
                 <View>
-                  <MovieCardOne
+                  <MovieDetailsCard
                     movie={selectedMovie}
                     movieResult={selectedMovieResult}
                   />
